+++ tests/test_search.py
@@ -1,6 +1,7 @@
 from tests import unittest, OrderedDict
 
 import jmespath
+import jmespath.functions
 
 
 class TestSearchOptions(unittest.TestCase):
@@ -10,3 +11,28 @@ class TestSearchOptions(unittest.TestCase):
             {'c': 'c', 'b': 'b', 'a': 'a', 'd': 'd'},
             options=jmespath.Options(dict_cls=OrderedDict))
         self.assertEqual(result, ['a', 'b', 'c'])
+
+    def test_can_provide_custom_functions(self):
+        class CustomFunctions(jmespath.functions.Functions):
+            @jmespath.functions.signature(
+                {'types': ['number']},
+                {'types': ['number']})
+            def _func_custom_add(self, x, y):
+                return x + y
+
+            @jmespath.functions.signature(
+                {'types': ['number']},
+                {'types': ['number']})
+            def _func_my_subtract(self, x, y):
+                return x - y
+
+
+        options = jmespath.Options(custom_functions=CustomFunctions())
+        self.assertEqual(
+            jmespath.search('custom_add(`1`, `2`)', {}, options=options),
+            3
+        )
+        self.assertEqual(
+            jmespath.search('my_subtract(`10`, `3`)', {}, options=options),
+            7
+        )